Add a stage1 configuration file
This configuration enable to use [stage1.io](http://stage1.io/) service.

Documentation about this configuration file:
http://help.stage1.io/article/customizing-a-build-with-the-build-yml-file/

Note that it could be use outside the sources of the project
(http://help.stage1.io/article/using-a-build-yml-without-adding-it-to-my-sources/).
However I prefer integrate it to verion control to dispatch any changes
with an update of the project installation.

This file is a copy of the original config suggestion of @ubermuda
(https://gist.github.com/ubermuda/bbac8b7f392508feb087) for openl10n
(https://twitter.com/ubermuda/status/481774170926510080).